tweaks + bugfixing

+++ source/nucledian/engine/graphics/shaders/sector.frag
@@ -22,11 +22,11 @@ flat in float tile_rotations_count;
 flat in float tile_rotation_increment;
 flat in vec2 texture_offset;
 
-layout(location = 0) out vec4  g_position;
-layout(location = 1) out vec4  g_normal;
-layout(location = 2) out vec3  g_stitched_normal;
-layout(location = 3) out vec4  g_albedo;
-layout(location = 4) out float g_sector;
+layout(location = 0) out vec4 g_position;
+layout(location = 1) out vec4 g_normal;
+layout(location = 2) out vec3 g_stitched_normal;
+layout(location = 3) out vec4 g_albedo;
+layout(location = 4) out uint g_sector;
 
 layout(binding = 0) uniform sampler2D game_atlas_sampler;
 layout(binding = 1) uniform sampler2D level_atlas_sampler;
